refactor: package for cleaning objects
