+++ src/main/java/fr/insee/lunatic/model/flat/Questionnaire.java
@@ -3,6 +3,7 @@ package fr.insee.lunatic.model.flat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import fr.insee.lunatic.model.flat.cleaning.CleaningType;
 import fr.insee.lunatic.model.flat.variable.VariableType;
 import lombok.Getter;
 import lombok.Setter;
